src: main: fix not resetting when playing recording
When recording finished and playback started, after it finishes first
play entities were not in place that they are started.

For example, let's say we recorded entity that was in (2, 3) that was
going 1 left, 2 up after it finishes playing first round, entity stays
in (1, 5) and moves again instead of returning to (2, 3).